xserver fontconfig mesa xkbcomp plink git update Jun 11 2018
xserver          commit 6300049a9a4b84789a2e8141f5cad17adf22185c
xkbcomp          commit 2abe23d23d3755335c10ff573e4e1f93b682e9d9
fontconfig       commit f5dd8512bdf9fd8e01c30ae36f593758b29385cf
mesa             commit 41642bdbca007035772fbfdc311f14daa5510d5d
putty            commit 93afcf02af7a68d736b487d9665aca741a3d5e0b

+++ fontconfig/test/run-test.sh
@@ -239,4 +239,19 @@ fi
 
 rm -rf $MyPWD/sysroot
 
+dotest "deleting .uuid file on empty dir"
+prep
+cp $FONT1 $FONT2 $FONTDIR
+$FCCACHE $FONTDIR
+sleep 1
+rm $FONTDIR/*pcf
+$FCCACHE $FONTDIR
+rmdir $FONTDIR > /dev/null 2>&1
+if [ $? != 0 ]; then
+  echo "*** Test failed: $TEST"
+  echo "$FONTDIR isn't empty"
+  ls -al $FONTDIR
+  exit 1
+fi
+
 rm -rf $FONTDIR $CACHEFILE $CACHEDIR $FONTCONFIG_FILE out
